requirements.txt
pytest
testcontainers[mongodb]
testcontainers[redis]
testcontainers[postgres]
sqlalchemy
psycopg2
factory_boy